Python: Don't allow pgen to regenerate grammar graminit.*
The Python tarball ships a Grammar file defining Python syntax.
C source files are generated from this as part of the build process.
First an executable called pgen is built, then pgen is run to
regenerate graminit.*.
In a cross-compile situation this goes terribly wrong, pgen is
compiled to the host architecture, meaning it's unable to run on
the build architecture.
Luckily the tarball ships pregenerated graminit.*, and we haven't modified
the Grammar file. So we can just copy across and force make to use them.