Python: Override PYTHON_FOR_BUILD correctly
If the environment variables $_PYTHON_PROJECT_BASE, $_PYTHON_HOST_PLATFORM, &
$PYTHONPATH are not set then when python's build script emits native values
for various config.h values and parameters. Needless to say we want the cross
build values.

The value we're using is the value set in configure.ac, after a search of $PATH
for a Python executable. Within cloudabi-ports that loop never succeeds,
because we don't include /usr/bin et al in `$PATH`