Generate importlib.h out of tree & ensure it is referenced there
Python requires two files (importlib.h & importlib_external.h) which are both
generated (from _bootstrap.py & _bootstrap_external.h). However, during an
out of tree build the regenerated copy is not referenced.

To fix this we add -IPython to the CPP flags. To make sure our regenerated
files are used we delete the versions shipped in the tarball.